gradually refactoring away from arango and full node to focus on etl-lite as the main data dependency
